Update Latin.properties
I chose extempore to use a lot of this vocabulary, such as ‘turba’ for unit, so it might be imperfect. The word for ‘turn’ also rather difficult since the Latin term is only really used in adverbial phrases and I want to use ‘aevum’ for technological eras. Something better (which isn’t Greek) will hopefully come to mind.